Changed ints to long, seems to have been at least once source of issues that showed up with HIP

+++ Source/PP_Spray/SprayDerive.cpp
@@ -58,7 +58,7 @@ SprayParticleContainer::computeDerivedVars(
       volfrac_fab = volfrac->array(pti);
     }
 #endif
-    amrex::ParallelFor(Np, [=] AMREX_GPU_DEVICE(int pid) noexcept {
+    amrex::ParallelFor(Np, [=] AMREX_GPU_DEVICE(Long pid) noexcept {
       const ParticleType& p = pstruct[pid];
       if (p.id() > 0) {
         RealVect lxc = (p.pos() - plo) * dxi;
